Legg til draft status på FortsettDialog også
Den var kun lagt til på Ny melding, og ble glemt å i fortsett dialog...
